+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v1.2.2
 * MenuWrapper 支持传入position，不一定要MouseEvent
 * `ContextMenuOption` 支持传入默认width
-
+* fix: inject `style` tag after `title` tag
 ## v1.2.1
 * MenuWrapper add destroy function to release DOM
 * fix: panel position error when scrollPage